+++ views/js/admin/shipment.js
@@ -101,9 +101,10 @@ $(document).ready(function () {
         var data = $(this).closest('form').serializeArray();
         var labelWindow;
         addButtonLoadingAnimation();
-        if (!is_label_download_option) {
+        if (!is_label_download_option && is_ps_above_177 && action != 'save') {
             labelWindow = window.open(loader_url, '_blank');
         }
+
         $.ajax(dpdAjaxShipmentsUrl, {
             method: 'POST',
             data: {
@@ -118,7 +119,6 @@ $(document).ready(function () {
                     response = JSON.parse(response);
                     removeButtonLoadingAnimation();
                     DPDhideError();
-
                     if (response.status == false) {
                         DPDshowError(response.message);
 
@@ -145,14 +145,23 @@ $(document).ready(function () {
                         hideShipmentError(response.id_dpd_shipment);
 
                     }
-
                     if ('save_and_print' === $clickedBtn.data('action')) {
+
                         var location = window.location +
                             '&print_label=1' +
                             '&id_dpd_shipment=' + encodeURIComponent(response.id_dpd_shipment);
 
+                        if (is_ps_above_177 && action === 'save_and_print' && print_and_save_label_url) {
+                            location = print_and_save_label_url
+                                .replace('orderId_', id_order);
+                        }
                         disableInputs(false);
-                        if (!is_label_download_option) {
+
+                        if (!is_label_download_option && !is_ps_above_177) {
+                            labelWindow = window.open(loader_url, '_blank');
+                        }
+
+                        if (!is_label_download_option && action !== 'save') {
                             labelWindow.location.href = location;
                         } else {
                             window.location.href = location;
@@ -242,10 +251,10 @@ $(document).ready(function () {
         var labelFormat = $('select[name="label_format"]').val();
         var labelPosition = $('input[name="label_position"]').val();
         var labelWindow;
-        if (!is_label_download_option) {
+
+        if (!is_label_download_option && is_ps_above_177) {
             labelWindow = window.open(loader_url, '_blank');
         }
-
         $.ajax(dpdAjaxShipmentsUrl, {
             method: 'POST',
             data: {
@@ -263,14 +272,28 @@ $(document).ready(function () {
                     DPDshowError(dpdMessages.unexpectedError);
                     return;
                 }
+
                 if (response.status) {
                     var location = window.location +
                         '&print_label=1' +
                         '&id_dpd_shipment=' + encodeURIComponent(shipmentId);
+                    if (is_ps_above_177 && action === 'print' && print_url) {
+                        location = print_url
+                            .replace('shipmentId_', shipmentId)
+                            .replace('labelFormat_', labelFormat)
+                            .replace('labelPosition_', labelPosition);
+                    } else if (is_ps_above_177 && action === 'save_and_print' && print_and_save_label_url) {
+                        location = print_and_save_label_url
+                            .replace('orderId_', id_order)
+                            .replace('isPrint_', true)
+                    }
 
                     disableInputs(false);
+                    if (!is_label_download_option && !is_ps_above_177) {
+                        labelWindow = window.open(loader_url, '_blank');
+                    }
 
-                    if (!is_label_download_option) {
+                    if (!is_label_download_option && action !== 'save') {
                         labelWindow.location.href = location;
                     } else {
                         window.location.href = location;